Fix silent call API using Login Hint
