+++ tests/Microsoft.Identity.Test.Unit.net45/PublicApiTests/PublicClientApplicationTests.cs
@@ -25,6 +25,21 @@
 //
 //------------------------------------------------------------------------------
 
+using Microsoft.Identity.Client;
+using Microsoft.Identity.Client.ApiConfig;
+using Microsoft.Identity.Client.AppConfig;
+using Microsoft.Identity.Client.Cache.Items;
+using Microsoft.Identity.Client.Core;
+using Microsoft.Identity.Client.OAuth2;
+using Microsoft.Identity.Client.TelemetryCore;
+using Microsoft.Identity.Client.UI;
+using Microsoft.Identity.Client.Utils;
+using Microsoft.Identity.Test.Common;
+using Microsoft.Identity.Test.Common.Core.Helpers;
+using Microsoft.Identity.Test.Common.Core.Mocks;
+using Microsoft.Identity.Test.Common.Mocks;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using NSubstitute;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -32,24 +47,6 @@ using System.Linq;
 using System.Net.Http;
 using System.Security;
 using System.Threading.Tasks;
-using Microsoft.Identity.Client;
-using Microsoft.Identity.Client.ApiConfig;
-using Microsoft.Identity.Client.AppConfig;
-using Microsoft.Identity.Client.Core;
-using Microsoft.Identity.Client.Cache;
-using Microsoft.Identity.Client.Cache.Items;
-using Microsoft.Identity.Client.Cache.Keys;
-using Microsoft.Identity.Client.Instance;
-using Microsoft.Identity.Client.OAuth2;
-using Microsoft.Identity.Client.TelemetryCore;
-using Microsoft.Identity.Client.UI;
-using Microsoft.Identity.Client.Utils;
-using Microsoft.Identity.Test.Common.Core.Helpers;
-using Microsoft.Identity.Test.Common.Core.Mocks;
-using Microsoft.Identity.Test.Common.Mocks;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using NSubstitute;
-using Microsoft.Identity.Test.Common;
 
 namespace Microsoft.Identity.Test.Unit.PublicApiTests
 {
@@ -617,506 +614,7 @@ namespace Microsoft.Identity.Test.Unit.PublicApiTests
             Assert.AreEqual(0, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
         }
 
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        public async Task AcquireTokenSilentScopeAndEmptyCacheTestAsync()
-        {
-            var receiver = new MyReceiver();
-            PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                        .WithAuthority(new Uri(ClientApplicationBase.DefaultAuthority), true)
-                                                                        .WithTelemetry(receiver.HandleTelemetryEvents)
-                                                                        .BuildConcrete();
 
-            try
-            {
-                AuthenticationResult result = await app.AcquireTokenSilentAsync(
-                                                  MsalTestConstants.Scope.ToArray(),
-                                                  new Account(
-                                                      MsalTestConstants.UserIdentifier,
-                                                      MsalTestConstants.DisplayableId,
-                                                      null)).ConfigureAwait(false);
-            }
-            catch (MsalUiRequiredException exc)
-            {
-                Assert.AreEqual(MsalUiRequiredException.NoTokensFoundError, exc.ErrorCode);
-            }
-
-            Assert.IsNotNull(
-                receiver.EventsReceived.Find(
-                    anEvent => // Expect finding such an event
-                        anEvent[EventBase.EventNameKey].EndsWith("api_event") && anEvent[ApiEvent.ApiIdKey] == "30" &&
-                        anEvent[ApiEvent.WasSuccessfulKey] == "false" &&
-                        anEvent[ApiEvent.ApiErrorCodeKey] == "no_tokens_found"));
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        public void AcquireTokenSilentScopeAndUserOverloadWithNoMatchingScopesInCacheTest()
-        {
-            // this test ensures that the API can
-            // get authority (if unique) from the cache entries where scope does not match.
-            // it should only happen for case where no authority is passed.
-
-            using (var httpManager = new MockHttpManager())
-            {
-                httpManager.AddInstanceDiscoveryMockHandler();
-
-                PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                            .WithAuthority(new Uri(ClientApplicationBase.DefaultAuthority), true)
-                                                                            .WithHttpManager(httpManager)
-                                                                            .BuildConcrete();
-
-                _tokenCacheHelper.PopulateCache(app.UserTokenCacheInternal.Accessor);
-                app.UserTokenCacheInternal.Accessor.DeleteAccessToken(
-                    new MsalAccessTokenCacheKey(
-                        MsalTestConstants.ProductionPrefNetworkEnvironment,
-                        MsalTestConstants.Utid,
-                        MsalTestConstants.UserIdentifier,
-                        MsalTestConstants.ClientId,
-                        MsalTestConstants.ScopeForAnotherResourceStr));
-
-                Task<AuthenticationResult> task = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.ScopeForAnotherResource.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null));
-
-                AuthenticationResult result = task.Result;
-                Assert.IsNotNull(result);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result.Account.Username);
-                Assert.AreEqual(MsalTestConstants.ScopeForAnotherResource.AsSingleString(), result.Scopes.AsSingleString());
-                Assert.AreEqual(2, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-            }
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        public void AcquireTokenSilentScopeAndUserOverloadDefaultAuthorityTest()
-        {
-            using (var httpManager = new MockHttpManager())
-            {
-                PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                            .WithAuthority(new Uri(MsalTestConstants.AuthorityTestTenant), true)
-                                                                            .WithHttpManager(httpManager)
-                                                                            .BuildConcrete();
-
-                _tokenCacheHelper.PopulateCache(app.UserTokenCacheInternal.Accessor);
-                app.UserTokenCacheInternal.Accessor.DeleteAccessToken(
-                    new MsalAccessTokenCacheKey(
-                        MsalTestConstants.ProductionPrefNetworkEnvironment,
-                        MsalTestConstants.Utid,
-                        MsalTestConstants.UserIdentifier,
-                        MsalTestConstants.ClientId,
-                        MsalTestConstants.ScopeForAnotherResourceStr));
-
-                httpManager.AddInstanceDiscoveryMockHandler();
-
-                Task<AuthenticationResult> task = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null));
-
-                AuthenticationResult result = task.Result;
-                Assert.IsNotNull(result);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.AsSingleString(), result.Scopes.AsSingleString());
-            }
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        public void AcquireTokenSilentScopeAndUserOverloadTenantSpecificAuthorityTest()
-        {
-            using (var httpManager = new MockHttpManager())
-            {
-                PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                            .WithAuthority(new Uri(MsalTestConstants.AuthorityGuestTenant), true)
-                                                                            .WithHttpManager(httpManager)
-                                                                            .BuildConcrete();
-                _tokenCacheHelper.PopulateCache(app.UserTokenCacheInternal.Accessor);
-                app.UserTokenCacheInternal.Accessor.DeleteAccessToken(
-                    new MsalAccessTokenCacheKey(
-                        MsalTestConstants.ProductionPrefNetworkEnvironment,
-                        MsalTestConstants.Utid,
-                        MsalTestConstants.UserIdentifier,
-                        MsalTestConstants.ClientId,
-                        MsalTestConstants.ScopeForAnotherResourceStr));
-
-                httpManager.AddInstanceDiscoveryMockHandler();
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityGuestTenant);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                Task<AuthenticationResult> task = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null));
-                AuthenticationResult result = task.Result;
-                Assert.IsNotNull(result);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.AsSingleString(), result.Scopes.AsSingleString());
-            }
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        public void AcquireTokenSilentCacheOnlyLookupTest()
-        {
-            var receiver = new MyReceiver();
-            using (var httpManager = new MockHttpManager())
-            {
-                PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                            .WithAuthority(new Uri(MsalTestConstants.AuthorityTestTenant), true)
-                                                                            .WithHttpManager(httpManager)
-                                                                            .WithTelemetry(receiver.HandleTelemetryEvents)
-                                                                            .BuildConcrete();
-                _tokenCacheHelper.PopulateCache(app.UserTokenCacheInternal.Accessor);
-
-                app.UserTokenCacheInternal.Accessor.DeleteAccessToken(new MsalAccessTokenCacheKey(
-                    MsalTestConstants.ProductionPrefNetworkEnvironment,
-                    MsalTestConstants.Utid,
-                    MsalTestConstants.UserIdentifier,
-                    MsalTestConstants.ClientId,
-                    MsalTestConstants.ScopeForAnotherResourceStr));
-
-                httpManager.AddInstanceDiscoveryMockHandler();
-
-                Task<AuthenticationResult> task = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    app.Authority,
-                    false);
-
-                AuthenticationResult result = task.Result;
-                Assert.IsNotNull(result);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.AsSingleString(), result.Scopes.AsSingleString());
-
-                Assert.AreEqual(2, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-
-                Assert.IsNotNull(receiver.EventsReceived.Find(anEvent =>  // Expect finding such an event
-                    anEvent[EventBase.EventNameKey].EndsWith("api_event") && anEvent[ApiEvent.WasSuccessfulKey] == "true"
-                    && anEvent[ApiEvent.ApiIdKey] == "31"));
-            }
-        }
-
-        [TestMethod]
-        [TestCategory("Regression")]
-        [WorkItem(695)] // Fix for https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/695
-        [TestCategory("PublicClientApplicationTests")]
-        public void AcquireTokenSilentForceRefreshTest()
-        {
-            using (var httpManager = new MockHttpManager())
-            {
-                PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                            .WithAuthority(new Uri(ClientApplicationBase.DefaultAuthority), true)
-                                                                            .WithHttpManager(httpManager)
-                                                                            .BuildConcrete();
-
-                _tokenCacheHelper.PopulateCacheWithOneAccessToken(app.UserTokenCacheInternal.Accessor);
-
-                httpManager.AddInstanceDiscoveryMockHandler();
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityUtidTenant);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                Task<AuthenticationResult> task = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    null,
-                    true);
-
-                AuthenticationResult result = task.Result;
-                Assert.IsNotNull(result);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.ToArray().AsSingleString(), result.Scopes.AsSingleString());
-
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-            }
-        }
-
-        [TestMethod]
-        [TestCategory("Regression")]
-        [WorkItem(695)] // Fix for https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/695
-        [TestCategory("PublicClientApplicationTests")]
-        public void AcquireTokenSilentForceRefreshMultipleTenantsTest()
-        {
-            using (var httpManager = new MockHttpManager())
-            {
-                PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                            .WithAuthority(new Uri(ClientApplicationBase.DefaultAuthority), true)
-                                                                            .WithHttpManager(httpManager)
-                                                                            .BuildConcrete();
-                _tokenCacheHelper.PopulateCacheWithOneAccessToken(app.UserTokenCacheInternal.Accessor);
-
-                httpManager.AddInstanceDiscoveryMockHandler();
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityCommonTenant);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                // ForceRefresh=true, so skip cache lookup of Access Token
-                // Use refresh token to acquire a new Access Token
-                Task<AuthenticationResult> task = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    MsalTestConstants.AuthorityCommonTenant,
-                    true);
-
-                AuthenticationResult result = task.Result;
-                Assert.IsNotNull(result);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.ToArray().AsSingleString(), result.Scopes.AsSingleString());
-
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityGuidTenant2);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                Task<AuthenticationResult> task2 = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    MsalTestConstants.AuthorityGuidTenant2,
-                    true);
-
-                // Same user, scopes, clientId, but different authority
-                // Should result in new AccessToken, but same refresh token
-                AuthenticationResult result2 = task2.Result;
-                Assert.IsNotNull(result2);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result2.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.ToArray().AsSingleString(), result2.Scopes.AsSingleString());
-
-                Assert.AreEqual(2, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityGuidTenant);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                // Same user, scopes, clientId, but different authority
-                // Should result in new AccessToken, but same refresh token
-                Task<AuthenticationResult> task3 = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    MsalTestConstants.AuthorityGuidTenant,
-                    true);
-
-                AuthenticationResult result3 = task3.Result;
-                Assert.IsNotNull(result3);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result3.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.ToArray().AsSingleString(), result3.Scopes.AsSingleString());
-
-                Assert.AreEqual(3, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-
-                // Use same authority as above, number of access tokens should remain constant
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                Task<AuthenticationResult> task4 = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    MsalTestConstants.AuthorityGuidTenant,
-                    true);
-
-                AuthenticationResult result4 = task4.Result;
-                Assert.IsNotNull(result4);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result4.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.ToArray().AsSingleString(), result4.Scopes.AsSingleString());
-
-                Assert.AreEqual(3, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-            }
-        }
-
-        [TestMethod]
-        [TestCategory("Regression")]
-        [WorkItem(695)] // Fix for https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/695
-        [TestCategory("PublicClientApplicationTests")]
-        public void AcquireTokenSilentForceRefreshFalseMultipleTenantsTest()
-        {
-            using (var httpManager = new MockHttpManager())
-            {
-                PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                            .WithAuthority(new Uri(ClientApplicationBase.DefaultAuthority), true)
-                                                                            .WithHttpManager(httpManager)
-                                                                            .BuildConcrete();
-
-                // PopulateCache() creates two access tokens
-                _tokenCacheHelper.PopulateCache(app.UserTokenCacheInternal.Accessor);
-
-                httpManager.AddInstanceDiscoveryMockHandler();
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityCommonTenant);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                Task<AuthenticationResult> task = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    MsalTestConstants.AuthorityCommonTenant,
-                    false);
-
-                AuthenticationResult result = task.Result;
-                Assert.IsNotNull(result);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.ToArray().AsSingleString(), result.Scopes.AsSingleString());
-
-                Assert.AreEqual(2, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityGuidTenant2);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                // Same user, scopes, clientId, but different authority
-                // Should result in new AccessToken, but same refresh token
-                Task<AuthenticationResult> task2 = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    MsalTestConstants.AuthorityGuidTenant2,
-                    false);
-
-                AuthenticationResult result2 = task2.Result;
-                Assert.IsNotNull(result2);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result2.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.ToArray().AsSingleString(), result2.Scopes.AsSingleString());
-
-                Assert.AreEqual(3, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityGuidTenant);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler()
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateSuccessTokenResponseMessage(
-                            MsalTestConstants.UniqueId,
-                            MsalTestConstants.DisplayableId,
-                            MsalTestConstants.Scope.ToArray())
-                    });
-
-                // Same user, scopes, clientId, but different authority
-                // Should result in new AccessToken, but same refresh token
-                Task<AuthenticationResult> task3 = app.AcquireTokenSilentAsync(
-                    MsalTestConstants.Scope.ToArray(),
-                    new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                    MsalTestConstants.AuthorityGuidTenant,
-                    false);
-
-                AuthenticationResult result3 = task3.Result;
-                Assert.IsNotNull(result3);
-                Assert.AreEqual(MsalTestConstants.DisplayableId, result3.Account.Username);
-                Assert.AreEqual(MsalTestConstants.Scope.ToArray().AsSingleString(), result3.Scopes.AsSingleString());
-
-                Assert.AreEqual(4, app.UserTokenCacheInternal.Accessor.AccessTokenCount);
-                Assert.AreEqual(1, app.UserTokenCacheInternal.Accessor.RefreshTokenCount);
-            }
-        }
-
-        [TestMethod]
-        [TestCategory("PublicClientApplicationTests")]
-        public void AcquireTokenSilentServiceErrorTest()
-        {
-            using (var httpManager = new MockHttpManager())
-            {
-                PublicClientApplication app = PublicClientApplicationBuilder.Create(MsalTestConstants.ClientId)
-                                                                            .WithAuthority(new Uri(ClientApplicationBase.DefaultAuthority), true)
-                                                                            .WithHttpManager(httpManager)
-                                                                            .BuildConcrete();
-                httpManager.AddInstanceDiscoveryMockHandler();
-                httpManager.AddMockHandlerForTenantEndpointDiscovery(MsalTestConstants.AuthorityCommonTenant);
-
-                //populate cache
-                _tokenCacheHelper.PopulateCache(app.UserTokenCacheInternal.Accessor);
-
-                httpManager.AddMockHandler(
-                    new MockHttpMessageHandler
-                    {
-                        ExpectedMethod = HttpMethod.Post,
-                        ResponseMessage = MockHelpers.CreateInvalidGrantTokenResponseMessage()
-                    });
-                try
-                {
-                    Task<AuthenticationResult> task = app.AcquireTokenSilentAsync(
-                        MsalTestConstants.CacheMissScope,
-                        new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                        app.Authority,
-                        false);
-                    AuthenticationResult result = task.Result;
-                    Assert.Fail("MsalUiRequiredException was expected");
-                }
-                catch (AggregateException ex)
-                {
-                    Assert.IsNotNull(ex.InnerException);
-                    Assert.IsTrue(ex.InnerException is MsalUiRequiredException);
-                    var msalExc = (MsalUiRequiredException)ex.InnerException;
-                    Assert.AreEqual(msalExc.ErrorCode, MsalUiRequiredException.InvalidGrantError);
-                }
-            }
-        }
 
         [TestMethod]
         [TestCategory("PublicClientApplicationTests")]
@@ -1529,10 +1027,13 @@ namespace Microsoft.Identity.Test.Unit.PublicApiTests
             CheckBuilderCommonMethods(deviceCodeBuilder);
 
             var silentBuilder = app.AcquireTokenSilent(MsalTestConstants.Scope, MsalTestConstants.User)
-               .WithForceRefresh(true)
-               .WithAccount(MsalTestConstants.User)
-               .WithLoginHint("loginhint");
+               .WithForceRefresh(true);
             CheckBuilderCommonMethods(silentBuilder);
+
+            silentBuilder = app.AcquireTokenSilent(MsalTestConstants.Scope, "upn@live.com")
+                .WithForceRefresh(true);
+            CheckBuilderCommonMethods(silentBuilder);
+
 
             var byRefreshTokenBuilder = ((IByRefreshToken)app).AcquireTokenByRefreshToken(MsalTestConstants.Scope, "refreshtoken")
                                   .WithRefreshToken("refreshtoken");
