+++ tests/Microsoft.Identity.Test.Unit.net45/RequestsTests/SilentRequestTests.cs
@@ -62,7 +62,6 @@ namespace Microsoft.Identity.Test.Unit.RequestsTests
         }
 
         [TestMethod]
-        [TestCategory("SilentRequestTests")]
         public void ConstructorTests()
         {
             using (var harness = new MockHttpTestHarness(MsalTestConstants.AuthorityHomeTenant))
@@ -82,7 +81,6 @@ namespace Microsoft.Identity.Test.Unit.RequestsTests
         }
 
         [TestMethod]
-        [TestCategory("SilentRequestTests")]
         public void ExpiredTokenRefreshFlowTest()
         {
             IDictionary<string, string> extraQueryParamsAndClaims =
@@ -93,9 +91,9 @@ namespace Microsoft.Identity.Test.Unit.RequestsTests
             {
                 _tokenCacheHelper.PopulateCache(harness.Cache.Accessor);
                 var parameters = harness.CreateRequestParams(
-                    harness.Cache, 
-                    null, 
-                    MsalTestConstants.ExtraQueryParams, 
+                    harness.Cache,
+                    null,
+                    MsalTestConstants.ExtraQueryParams,
                     MsalTestConstants.Claims);
                 var silentParameters = new AcquireTokenSilentParameters();
 
@@ -130,7 +128,6 @@ namespace Microsoft.Identity.Test.Unit.RequestsTests
         }
 
         [TestMethod]
-        [TestCategory("SilentRequestTests")]
         public void SilentRefreshFailedNullCacheTest()
         {
             using (var harness = new MockHttpTestHarness(MsalTestConstants.AuthorityHomeTenant))
@@ -163,7 +160,6 @@ namespace Microsoft.Identity.Test.Unit.RequestsTests
         }
 
         [TestMethod]
-        [TestCategory("SilentRequestTests")]
         public void SilentRefreshFailedNoCacheItemFoundTest()
         {
             using (var harness = new MockHttpTestHarness(MsalTestConstants.AuthorityHomeTenant))
@@ -219,9 +215,9 @@ namespace Microsoft.Identity.Test.Unit.RequestsTests
             }
 
             public AuthenticationRequestParameters CreateRequestParams(
-                ITokenCacheInternal cache, 
-                SortedSet<string> scopes, 
-                IDictionary<string, string> extraQueryParams = null, 
+                ITokenCacheInternal cache,
+                SortedSet<string> scopes,
+                IDictionary<string, string> extraQueryParams = null,
                 string claims = null)
             {
                 var commonParameters = new AcquireTokenCommonParameters
@@ -239,7 +235,7 @@ namespace Microsoft.Identity.Test.Unit.RequestsTests
                     RequestContext.CreateForTest(ServiceBundle))
                 {
                     Account = new Account(MsalTestConstants.UserIdentifier, MsalTestConstants.DisplayableId, null),
-                   
+
                 };
                 return parameters;
             }
